Colocado validacao do carrinho para bloquear a persistencia se não passar na validação

+++ src/services/NSE.Carrinho.API/Controllers/CarrinhoController.cs
@@ -6,6 +6,7 @@ using NSE.Carrinho.API.Models;
 using NSE.WebAPI.Core.Controllers;
 using NSE.WebAPI.Core.Usuario;
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace NSE.Carrinho.API.Controllers
@@ -62,6 +63,9 @@ namespace NSE.Carrinho.API.Controllers
 
 			carrinho.AtualizarUnidades(itemCarrinho, item.Quantidade);
 
+			ValidarCarrinho(carrinho);
+			if (!OperacaoValida()) return CustomResponse();
+
 			_context.CarrinhoItens.Update(itemCarrinho);
 			_context.CarrinhoClientes.Update(carrinho);
 
@@ -78,6 +82,9 @@ namespace NSE.Carrinho.API.Controllers
 			if (itemCarrinho == null) return CustomResponse();
 
 			carrinho.RemoverItem(itemCarrinho);
+
+			ValidarCarrinho(carrinho);
+			if (!OperacaoValida()) return CustomResponse();
 
 			_context.CarrinhoItens.Remove(itemCarrinho);
 			_context.CarrinhoClientes.Update(carrinho);
@@ -149,6 +156,14 @@ namespace NSE.Carrinho.API.Controllers
 		{
 			var result = await _context.SaveChangesAsync();
 			if (result <= 0) AdicionarErroProcessamento("Não foi possível persistir os dados no banco");
+		}
+
+		private bool ValidarCarrinho(CarrinhoCliente carrinho)
+		{
+			if (carrinho.EhValido()) return true;
+
+			carrinho.ValidationResult.Errors.ToList().ForEach(e => AdicionarErroProcessamento(e.ErrorMessage));
+			return false;
 		}
 	}
 }
